fix(council): update QualityOracle to match simplified interface
- Changed verifyScore to view function (no state modification)
- Changed score type from uint8 to uint256 for consistency
- Renamed parameters: submitter→proposer, timestamp→attestationTimestamp
- Removed checkAttestation function (redundant with view verifyScore)
- Removed replay protection (handled by Council contract)
- Updated tests to match new interface

+++ apps/council/src/proposal-assistant.ts
@@ -276,7 +276,7 @@ Return JSON: {"title":"...","summary":"...","description":"..."}`;
 
     // Build the message hash (must match QualityOracle.sol)
     const messageHash = solidityPackedKeccak256(
-      ['string', 'bytes32', 'uint8', 'uint256', 'address', 'uint256'],
+      ['string', 'bytes32', 'uint256', 'uint256', 'address', 'uint256'],
       ['JejuQualityAttestation', contentHash, score, timestamp, submitterAddress, chainId]
     );
